+++ actors/paych/src/types.rs
@@ -14,14 +14,14 @@ use fvm_shared::MethodNum;
 use super::Merge;
 
 /// Maximum number of lanes in a channel
-pub const MAX_LANE: i64 = std::i64::MAX;
+pub const MAX_LANE: u64 = std::i64::MAX as u64;
 
 pub const SETTLE_DELAY: ChainEpoch = EPOCHS_IN_HOUR * 12;
 
 // Maximum byte length of a secret that can be submitted with a payment channel update.
 pub const MAX_SECRET_SIZE: usize = 256;
 
-pub const LANE_STATES_AMT_BITWIDTH: usize = 3;
+pub const LANE_STATES_AMT_BITWIDTH: u32 = 3;
 /// Constructor parameters for payment channel actor
 #[derive(Serialize_tuple, Deserialize_tuple)]
 pub struct ConstructorParams {
@@ -46,7 +46,7 @@ pub struct SignedVoucher {
     /// (optional) Specified by `from` to add a verification method to the voucher
     pub extra: Option<ModVerifyParams>,
     /// Specifies which lane the Voucher merges into (will be created if does not exist)
-    pub lane: usize,
+    pub lane: u64,
     /// Set by `from` to prevent redemption of stale vouchers on a lane
     pub nonce: u64,
     /// Amount voucher can be redeemed for
@@ -73,7 +73,7 @@ impl SignedVoucher {
             #[serde(with = "serde_bytes")]
             pub secret_pre_image: &'a [u8],
             pub extra: &'a Option<ModVerifyParams>,
-            pub lane: usize,
+            pub lane: u64,
             pub nonce: u64,
             #[serde(with = "bigint_ser")]
             pub amount: &'a BigInt,
